refactor: move otlp contracts to dedicated folder
